First idea of AAdam ... try the new version
This very first idea actually works well. However, the results are sometimes mitigate since the sign of the gradient may result to a very bad behavior even for simple convex problems such as logistic regression.
To cope with this issue, we move out the sign and use its variation as an heuristic. See AAdam and AAdam2.